pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bible-data-pipeline"
version = "1.0.0"
description = "🕊️ A production-grade data engineering pipeline for Biblical text analytics"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your@email.com"},
]
keywords = ["data-engineering", "etl", "bible", "nlp", "analytics", "duckdb", "gcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "duckdb>=1.1.0",
    "pydantic>=2.7.0",
    "textblob>=0.18.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sentry-sdk[fastapi]>=2.20.0",
    "google-generativeai>=0.8.0",
    "defusedxml>=0.7.0",
]

[project.optional-dependencies]
gcp = [
    "google-cloud-storage>=2.16.0",
    "google-cloud-bigquery>=3.20.0",
    "google-cloud-texttospeech>=2.16.0",
]
dashboard = [
    "streamlit>=1.37.0",
    "plotly>=5.22.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
all = ["bible-data-pipeline[gcp,dashboard,dev]"]

[project.scripts]
bible-pipeline = "src.cli:app"

[project.urls]
Homepage = "https://github.com/DavidKGBR/the-bible"
Repository = "https://github.com/DavidKGBR/the-bible"
Issues = "https://github.com/DavidKGBR/the-bible/issues"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PT", "RET", "SIM"]
ignore = ["S101", "S603", "S607", "S608", "S108", "RET504", "B027", "B905", "PT018", "SIM117", "C401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html -m 'not integration and not slow'"
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
disable_error_code = ["prop-decorator"]
